+++ CHANGELOG.md
@@ -37,3 +37,4 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - add a retryable flag to distinguish recoverable errors
+- improve retry logic in S7ClientSession